--- parent_parnter_creation_restrict/readme/DESCRIPTION.md
@@ -1,1 +1,0 @@
-This module restricts the creation and editing of parent partners, except for certain authorized user groups.
