+++ partner_group_partner_admin/__manifest__.py
@@ -1,12 +1,13 @@
 # Copyright 2024 Quartile Limited
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 {
-    "name": "Parent Partner Creation Restrict",
+    "name": "Partner Group Partner Admin",
     "version": "16.0.1.0.0",
     "author": "Quartile Limited",
     "website": "https://www.quartile.co",
     "category": "Contact",
     "license": "AGPL-3",
-    "depends": ["sale", "purchase"],
+    "depends": ["base"],
+    "data": ["security/partner_security.xml"],
     "installable": True,
 }
