+++ partner_group_partner_admin/readme/DESCRIPTION.md
@@ -1,0 +1,3 @@
+This module adds the partner administrator group to do the following:
+
+* Limits the creation and edition of commercial partners to the partner administrator.
