+++ partner_group_partner_admin/security/partner_security.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <record id="group_partner_admin" model="res.groups">
+        <field name="name">Partner Administrator</field>
+        <field name="category_id" ref="base.module_category_usability" />
+    </record>
+</odoo>
