+++ partner_group_partner_admin/static/description/index.html
@@ -3,7 +3,7 @@
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
 <meta name="generator" content="Docutils: https://docutils.sourceforge.io/" />
-<title>Parent Partner Creation Restrict</title>
+<title>Partner Group Partner Admin</title>
 <style type="text/css">
 
 /*
@@ -359,8 +359,8 @@ ul.auto-toc {
 </style>
 </head>
 <body>
-<div class="document" id="parent-partner-creation-restrict">
-<h1 class="title">Parent Partner Creation Restrict</h1>
+<div class="document" id="partner-group-partner-admin">
+<h1 class="title">Partner Group Partner Admin</h1>
 
 <!-- !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 !! This file is generated by oca-gen-addon-readme !!
@@ -368,9 +368,12 @@ ul.auto-toc {
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 !! source digest: sha256:2b07697c0400462fedd824b239b455637c4776a7027c7a773aff733a635fec46
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! -->
-<p><a class="reference external image-reference" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external image-reference" href="http://www.gnu.org/licenses/agpl-3.0-standalone.html"><img alt="License: AGPL-3" src="https://img.shields.io/badge/licence-AGPL--3-blue.png" /></a> <a class="reference external image-reference" href="https://github.com/qrtl/thc-custom/tree/16.0/parent_parnter_creation_restrict"><img alt="qrtl/thc-custom" src="https://img.shields.io/badge/github-qrtl%2Fthc--custom-lightgray.png?logo=github" /></a></p>
-<p>This module restricts the creation and editing of parent partners,
-except for certain authorized user groups.</p>
+<p><a class="reference external image-reference" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external image-reference" href="http://www.gnu.org/licenses/agpl-3.0-standalone.html"><img alt="License: AGPL-3" src="https://img.shields.io/badge/licence-AGPL--3-blue.png" /></a> <a class="reference external image-reference" href="https://github.com/qrtl/thc-custom/tree/16.0/partner_group_partner_admin"><img alt="qrtl/thc-custom" src="https://img.shields.io/badge/github-qrtl%2Fthc--custom-lightgray.png?logo=github" /></a></p>
+<p>This module adds the partner administrator group to do the following:</p>
+<ul class="simple">
+<li>Limits the creation and edition of commercial partners to the partner
+administrator.</li>
+</ul>
 <p><strong>Table of contents</strong></p>
 <div class="contents local topic" id="contents">
 <ul class="simple">
@@ -387,7 +390,7 @@ except for certain authorized user groups.</p>
 <p>Bugs are tracked on <a class="reference external" href="https://github.com/qrtl/thc-custom/issues">GitHub Issues</a>.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us to smash it by providing a detailed and welcomed
-<a class="reference external" href="https://github.com/qrtl/thc-custom/issues/new?body=module:%20parent_parnter_creation_restrict%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**">feedback</a>.</p>
+<a class="reference external" href="https://github.com/qrtl/thc-custom/issues/new?body=module:%20partner_group_partner_admin%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**">feedback</a>.</p>
 <p>Do not contact contributors directly about support or help with technical issues.</p>
 </div>
 <div class="section" id="credits">
@@ -400,7 +403,7 @@ If you spotted it first, help us to smash it by providing a detailed and welcome
 </div>
 <div class="section" id="maintainers">
 <h2><a class="toc-backref" href="#toc-entry-4">Maintainers</a></h2>
-<p>This module is part of the <a class="reference external" href="https://github.com/qrtl/thc-custom/tree/16.0/parent_parnter_creation_restrict">qrtl/thc-custom</a> project on GitHub.</p>
+<p>This module is part of the <a class="reference external" href="https://github.com/qrtl/thc-custom/tree/16.0/partner_group_partner_admin">qrtl/thc-custom</a> project on GitHub.</p>
 <p>You are welcome to contribute.</p>
 </div>
 </div>
