--- setup/parent_parnter_creation_restrict/odoo/addons/parent_parnter_creation_restrict
@@ -1,1 +1,0 @@
-../../../../parent_parnter_creation_restrict
