+++ setup/partner_group_partner_admin/odoo/addons/partner_group_partner_admin
@@ -1,0 +1,1 @@
+../../../../partner_group_partner_admin
